+++ cardano-api/src/Cardano/Api.hs
@@ -334,10 +334,14 @@ module Cardano.Api (
     -- * Transaction metadata
     -- | Embedding additional structured data within transactions.
     TxMetadata(..),
+    AsTxMetadata(..),
 
     -- ** Constructing metadata
     TxMetadataValue(..),
     makeTransactionMetadata,
+    mergeTransactionMetadata,
+    metaTextChunks,
+    metaBytesChunks,
 
     -- ** Validating metadata
     validateTxMetadata,