+++ cardano-cli/src/Cardano/CLI/Shelley/Run/Governance.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+
 module Cardano.CLI.Shelley.Run.Governance
   ( ShelleyGovernanceCmdError
   , renderShelleyGovernanceError
@@ -5,14 +8,23 @@ module Cardano.CLI.Shelley.Run.Governance
   ) where
 
 import           Control.Monad (unless, when)
+import           Control.Monad.IO.Class (liftIO)
 import           Control.Monad.Trans.Except (ExceptT)
 import           Control.Monad.Trans.Except.Extra (firstExceptT, handleIOExceptT, left, newExceptT,
                    onLeft)
 import           Data.Aeson (eitherDecode)
+import qualified Data.ByteString.Char8 as BSC
 import qualified Data.ByteString.Lazy as LB
 import           Data.Function ((&))
+import qualified Data.List as List
+import           Data.String (fromString)
 import           Data.Text (Text)
 import qualified Data.Text as Text
+import qualified Data.Text.Encoding as Text
+import qualified Data.Text.IO as Text
+import qualified Data.Text.Read as Text
+import           Formatting (build, sformat)
+import           System.IO (stderr, stdin, stdout)
 
 import           Cardano.Api
 import           Cardano.Api.Shelley
@@ -20,10 +32,16 @@ import           Cardano.Api.Shelley
 import           Cardano.CLI.Shelley.Key (VerificationKeyOrHashOrFile,
                    readVerificationKeyOrHashOrFile, readVerificationKeyOrHashOrTextEnvFile)
 import           Cardano.CLI.Shelley.Parsers
+import           Cardano.CLI.Shelley.Run.Key (SomeSigningKey (..), readSigningKeyFile)
+import           Cardano.CLI.Shelley.Run.Read (MetadataError, readFileTxMetadata,
+                   renderMetadataError)
 import           Cardano.CLI.Types
 
+import           Cardano.Binary (DecoderError)
+import           Cardano.Ledger.Alonzo.Scripts (CostModels (..))
+import           Cardano.Ledger.Crypto (StandardCrypto)
+import           Cardano.Ledger.Keys (SignKeyDSIGN, SignKeyVRF)
 import qualified Cardano.Ledger.Shelley.TxBody as Shelley
-
 
 data ShelleyGovernanceCmdError
   = ShelleyGovernanceCmdTextEnvReadError !(FileError TextEnvelopeError)
@@ -39,6 +57,16 @@ data ShelleyGovernanceCmdError
       -- ^ Number of reward amounts
   | ShelleyGovernanceCmdCostModelsJsonDecodeErr !FilePath !Text
   | ShelleyGovernanceCmdEmptyCostModel !FilePath
+  | ShelleyGovernanceCmdUnexpectedKeyType
+      ![TextEnvelopeType]
+      -- ^ Expected key types
+  | ShelleyGovernanceCmdPollOutOfBoundAnswer
+      !Int
+      -- ^ Maximum answer index
+  | ShelleyGovernanceCmdPollInvalidChoice
+  | ShelleyGovernanceCmdMetadataError !MetadataError
+  | ShelleyGovernanceCmdDecoderError !DecoderError
+  | ShelleyGovernanceCmdVerifyPollError !GovernancePollError
   deriving Show
 
 renderShelleyGovernanceError :: ShelleyGovernanceCmdError -> Text
@@ -61,7 +89,19 @@ renderShelleyGovernanceError err =
       "The decoded cost model was empty at: " <> Text.pack fp
     ShelleyGovernanceCmdCostModelReadError err' ->
       "Error reading the cost model: " <> Text.pack (displayError err')
-
+    ShelleyGovernanceCmdUnexpectedKeyType expected ->
+      "Unexpected poll key type; expected one of: "
+      <> Text.intercalate ", " (textShow <$> expected)
+    ShelleyGovernanceCmdPollOutOfBoundAnswer nMax ->
+      "Poll answer out of bounds. Choices are between 0 and " <> textShow nMax
+    ShelleyGovernanceCmdPollInvalidChoice ->
+      "Invalid choice. Please choose from the available answers."
+    ShelleyGovernanceCmdMetadataError metadataError ->
+      renderMetadataError metadataError
+    ShelleyGovernanceCmdDecoderError decoderError ->
+      "Unable to decode metadata: " <> sformat build decoderError
+    ShelleyGovernanceCmdVerifyPollError pollError ->
+      renderGovernancePollError pollError
 
 runGovernanceCmd :: GovernanceCmd -> ExceptT ShelleyGovernanceCmdError IO ()
 runGovernanceCmd (GovernanceMIRPayStakeAddressesCertificate mirpot vKeys rewards out) =
@@ -72,6 +112,12 @@ runGovernanceCmd (GovernanceGenesisKeyDelegationCertificate genVk genDelegVk vrf
   runGovernanceGenesisKeyDelegationCertificate genVk genDelegVk vrfVk out
 runGovernanceCmd (GovernanceUpdateProposal out eNo genVKeys ppUp mCostModelFp) =
   runGovernanceUpdateProposal out eNo genVKeys ppUp mCostModelFp
+runGovernanceCmd (GovernanceCreatePoll prompt choices nonce out) =
+  runGovernanceCreatePoll prompt choices nonce out
+runGovernanceCmd (GovernanceAnswerPoll poll sk ix) =
+  runGovernanceAnswerPoll poll sk ix
+runGovernanceCmd (GovernanceVerifyPoll poll metadata) =
+  runGovernanceVerifyPoll poll metadata
 
 runGovernanceMIRCertificatePayStakeAddrs
   :: Shelley.MIRPot
@@ -180,3 +226,158 @@ runGovernanceUpdateProposal (OutputFile upFile) eNo genVerKeyFiles upPprams mCos
   firstExceptT ShelleyGovernanceCmdTextEnvWriteError . newExceptT
     $ writeLazyByteStringFile upFile $ textEnvelopeToJSON Nothing upProp
 
+runGovernanceCreatePoll
+  :: Text
+  -> [Text]
+  -> Maybe Word
+  -> OutputFile
+  -> ExceptT ShelleyGovernanceCmdError IO ()
+runGovernanceCreatePoll govPollQuestion govPollAnswers govPollNonce out = do
+  let poll = GovernancePoll{ govPollQuestion, govPollAnswers, govPollNonce }
+
+  let description = fromString $ "An on-chain poll for SPOs: " <> Text.unpack govPollQuestion
+  firstExceptT ShelleyGovernanceCmdTextEnvWriteError . newExceptT $
+    writeFileTextEnvelope (unOutputFile out) (Just description) poll
+
+  let metadata = asTxMetadata poll
+        & metadataToJson TxMetadataJsonDetailedSchema
+
+  let outPath = unOutputFile out
+        & Text.encodeUtf8 . Text.pack
+
+  liftIO $ do
+    BSC.hPutStrLn stderr $ mconcat
+      [ "Poll created successfully.\n"
+      , "Please submit a transaction using the resulting metadata.\n"
+      ]
+    BSC.hPutStrLn stdout (prettyPrintJSON metadata)
+    BSC.hPutStrLn stderr $ mconcat
+      [ "\n"
+      , "Hint (1): Use '--json-metadata-detailed-schema' and '--metadata-json-file' "
+      , "from the build or build-raw commands.\n"
+      , "Hint (2): You can redirect the standard output of this command to a JSON "
+      , "file to capture metadata.\n\n"
+      , "Note: A serialized version of the poll suitable for sharing with "
+      , "participants has been generated at '" <> outPath <> "'."
+      ]
+
+runGovernanceAnswerPoll
+  :: FilePath
+  -> SigningKeyFile
+    -- ^ VRF or Ed25519 cold key
+  -> Maybe Word
+    -- ^ Answer index
+  -> ExceptT ShelleyGovernanceCmdError IO ()
+runGovernanceAnswerPoll pollFile skFile maybeChoice = do
+  poll <- firstExceptT ShelleyGovernanceCmdTextEnvReadError . newExceptT $
+    readFileTextEnvelope AsGovernancePoll pollFile
+
+  credentials <- readVRFOrColdSigningKeyFile skFile
+
+  choice <- case maybeChoice of
+    Nothing -> do
+      askInteractively poll
+    Just ix -> do
+      validateChoice poll ix
+      liftIO $ BSC.hPutStrLn stderr $ Text.encodeUtf8 $ Text.intercalate "\n"
+        [ govPollQuestion poll
+        , "→ " <> (govPollAnswers poll !! fromIntegral ix)
+        , ""
+        ]
+      pure ix
+
+  let pollAnswer = GovernancePollAnswer
+        { govAnsPoll = hashGovernancePoll poll
+        , govAnsChoice = choice
+        }
+  let witness = pollAnswer `signPollAnswerWith` credentials
+
+  let metadata =
+        mergeTransactionMetadata
+          ( \l r -> case (l, r) of
+              (TxMetaMap xs, TxMetaMap ys) -> TxMetaMap (xs <> ys)
+              _ -> error "unreachable"
+          )
+          (asTxMetadata pollAnswer)
+          (asTxMetadata witness)
+        & metadataToJson TxMetadataJsonDetailedSchema
+
+  liftIO $ do
+    BSC.hPutStrLn stderr $ mconcat
+      [ "Poll answer created successfully.\n"
+      , "Please submit a transaction using the resulting metadata.\n"
+      ]
+    BSC.hPutStrLn stdout (prettyPrintJSON metadata)
+    BSC.hPutStrLn stderr $ mconcat
+      [ "\n"
+      , "Hint (1): Use '--json-metadata-detailed-schema' and '--metadata-json-file' "
+      , "from the build or build-raw commands.\n"
+      , "Hint (2): You can redirect the standard output of this command to a JSON "
+      , "file to capture metadata."
+      ]
+ where
+  readVRFOrColdSigningKeyFile
+    :: SigningKeyFile
+    -> ExceptT
+         ShelleyGovernanceCmdError
+         IO
+         (Either (SignKeyVRF StandardCrypto) (SignKeyDSIGN StandardCrypto))
+  readVRFOrColdSigningKeyFile filepath = do
+    someSk <- firstExceptT ShelleyGovernanceCmdKeyReadError $
+      readSigningKeyFile filepath
+    case someSk of
+      AVrfSigningKey (VrfSigningKey sk) ->
+        pure (Left sk)
+      AStakePoolSigningKey (StakePoolSigningKey sk) ->
+        pure (Right sk)
+      _anythingElse ->
+        left $ ShelleyGovernanceCmdUnexpectedKeyType
+          [ textEnvelopeType (AsSigningKey AsVrfKey)
+          , textEnvelopeType (AsSigningKey AsStakePoolKey)
+          ]
+
+  validateChoice :: GovernancePoll -> Word -> ExceptT ShelleyGovernanceCmdError IO ()
+  validateChoice GovernancePoll{govPollAnswers} ix = do
+    let maxAnswerIndex = length govPollAnswers - 1
+    when (fromIntegral ix > maxAnswerIndex) $ left $
+      ShelleyGovernanceCmdPollOutOfBoundAnswer maxAnswerIndex
+
+  askInteractively :: GovernancePoll -> ExceptT ShelleyGovernanceCmdError IO Word
+  askInteractively poll@GovernancePoll{govPollQuestion, govPollAnswers} = do
+    liftIO $ BSC.hPutStrLn stderr $ Text.encodeUtf8 $ Text.intercalate "\n"
+      ( govPollQuestion
+      : [ "[" <> textShow ix <> "] " <> answer
+        | (ix :: Int, answer) <- zip [0..] govPollAnswers
+        ]
+      )
+    liftIO $ BSC.hPutStrLn stderr ""
+    liftIO $ BSC.hPutStr stderr "Please indicate an answer (by index): "
+    txt <- liftIO $ Text.hGetLine stdin
+    liftIO $ BSC.hPutStrLn stderr ""
+    case Text.decimal txt of
+      Right (choice, rest) | Text.null rest ->
+        choice <$ validateChoice poll choice
+      _ ->
+        left ShelleyGovernanceCmdPollInvalidChoice
+
+runGovernanceVerifyPoll
+  :: FilePath
+  -> FilePath
+  -> ExceptT ShelleyGovernanceCmdError IO ()
+runGovernanceVerifyPoll pollFile metadataFile = do
+  poll <- firstExceptT ShelleyGovernanceCmdTextEnvReadError . newExceptT $
+    readFileTextEnvelope AsGovernancePoll pollFile
+
+  metadata <- firstExceptT ShelleyGovernanceCmdMetadataError $
+    readFileTxMetadata TxMetadataJsonDetailedSchema (MetadataFileJSON metadataFile)
+
+  answer <- firstExceptT ShelleyGovernanceCmdDecoderError . newExceptT $ pure $
+    deserialiseFromCBOR AsGovernancePollAnswer (serialiseToCBOR metadata)
+
+  witness <- firstExceptT ShelleyGovernanceCmdDecoderError . newExceptT $ pure $
+    deserialiseFromCBOR AsGovernancePollWitness (serialiseToCBOR metadata)
+
+  firstExceptT ShelleyGovernanceCmdVerifyPollError . newExceptT $ pure $
+    verifyPollAnswer poll answer witness
+
+  liftIO $ BSC.hPutStrLn stderr "Ok."