+++ cardano-cli/test/cardano-cli-golden.hs
@@ -26,5 +26,6 @@ main = do
     , Test.Golden.Shelley.metadataTests
     , Test.Golden.Shelley.multiSigTests
     , Test.Golden.Shelley.txTests
+    , Test.Golden.Shelley.governancePollTests
     , Test.Golden.TxView.txViewTests
     ]